+++ RELEASE_NOTES.md
@@ -4,9 +4,14 @@
 
 ---
 
-Changes (live document)
-- Added messages.dbggreet logic to view-home.handlebars for msg debugging.
 
+
+## **1.0.3** (Oct 17 2020)
+- added support for "filelist" dropdown selector for choosing a file of specific type (such as 'png') from a specific Asset -folder (such as '/media/image/logo/')
+- added messages.dbggreet logic to view-home.handlebars for msg debugging.
+- improved "continue" button logic in UI and added "steps" parameter to template definition
+- moved templates/empty.html from a file to internal route
+- implemented duplicate and rename rundown -buttons in the rundown list view
 
 ## **1.0.2** (Sept 21 2020)
 - added support for multiline "textarea" fields